+++ src/scietex/logging/__init__.py
@@ -97,7 +97,7 @@ behaviors.
 
 """
 
-__version__ = "1.0.0"
+__version__ = "1.1.0"
 
 from .async_logging_handler import AsyncLoggingHandler
 from .basic_handler import AsyncBaseHandler